Do not apply hidden-activation to output units!
This change hopefully fixes a recent break in RNN trainability.

It also entails changing the signature of the add_recurrent_layer method
so that it no longer returns the computed activations of the layer.
Instead, both the preactivations and the hidden activations are stored
in .hiddens, and the caller can fiddle with those as needed after the
call.