Replace norm scaling with elementwise clipping.
This ought to work better in terms of making clipping apply more fairly
across all different sorts of parameter sizes. E.g. a bias vector is
probably going to have a small number of values in it, while a recurrent
hidden-hidden matrix will have a large number ... if we scale the norms
of these gradients, the individual step sizes in a scaled hidden-hidden
matrix will be much smaller than the sizes for the bias vector. (But
maybe that's good??) If we clip the gradients elementwise, we're
basically putting a cap on the delta that each individual parameter is
allowed to experience on a given update.