Use strings, not numbers, to specify layer inputs.
The inputs quantity for a layer used to be used to initialize the
layer's parameters, which required the size of the input/output
quantities to be known at layer creation time. Using numbers for layer
inputs was mostly ambiguous---multiple layers in a network might have
the same size.

Having earlier introduced a two-pass graph binding process for layers,
now layer inputs are treated as strings naming layer outputs. These
names are resolved at binding time, not at creation time.