Allow symbolic initial recurrent state.
Traditionally the initial state for a recurrent network is set to all
zeros, the assumption being that recurrent models will run for
sufficiently many time steps to amortize the effect of the initial
state. This change permits the initial state vectors to come from a
network output instead, which should allow for single-time-step model
stepping.

Contributes to issue #81.