Change handling of shape/size for layers.
This change enables full shape inference by layers that need
it (especially convolution layers)! The idea is that at creation time
each layer basically only knows the names of the things it is to be
connected to, and then at binding time the shapes of inputs and outputs
are inferred from the full set of graph layers.

This all might be better installed in the "connect" method, but it works
for the time being.

Addresses issue #126.